requirements.txt
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cssutils==2.11.1
curl_cffi==0.13.0
cycler==0.12.1
et_xmlfile==2.0.0
finnhub-python==2.4.24
fonttools==4.59.2
fpdf==1.7.2
frozendict==2.4.6
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
joblib==1.5.2
kiwisolver==1.4.9
loguru==0.7.3
lxml==5.4.0
matplotlib==3.10.6
more-itertools==10.8.0
multitasking==0.0.12
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
protobuf==6.32.1
psycopg2-binary==2.9.10
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.4
pyrate-limiter==3.9.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sec-downloader==0.11.2
sec-edgar-downloader==5.0.3
sec-parser==0.58.1
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.43
tabulate==0.9.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
win32_setctime==1.2.0
xlsxwriter==3.2.8
xxhash==3.5.0
yfinance==0.2.65
